requirements.txt
fastapi
uvicorn
python-dotenv
groq
python-multipart
jinja2
pydantic
pydantic-settings
aiofiles
pytest
httpx